requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
pymongo==4.5.0
python-dotenv==1.0.0
requests==2.31.0
asyncio==3.4.3
aiohttp==3.8.6
python-youtube==0.8.2
anthropic>=0.25.0
openai==0.28.1
nest-asyncio>=1.5.0
google-api-python-client==2.100.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.1
Pillow==10.0.1
moviepy==1.0.3
pydub==0.25.1
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
webdriver-manager==4.0.1
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2023.3
colorama==0.4.6
tqdm==4.66.1
numpy==1.24.4
pandas==2.1.1
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0
dash==2.14.1
dash-bootstrap-components==1.5.0
instagrapi==2.0.0
replicate>=0.15.0
opencv-python-headless==4.8.1.78 